feat : 권한 검증 추가 (조회, 삭제는 마스터 유저만 접근 가능)

+++ client/alarm/src/main/java/com/logistic/client/alarm/presentation/controller/SlackController.java
@@ -5,6 +5,8 @@ import com.logistic.client.alarm.application.dto.SlackResponseDto;
 import com.logistic.client.alarm.application.service.SlackApplicationService;
 import com.logistic.client.alarm.presentation.request.SlackRequestDto;
 import com.logistic.client.alarm.presentation.request.SlackSearchDto;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.ws.rs.ForbiddenException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -25,19 +27,34 @@ public class SlackController {
     }
 
     @GetMapping("/{slackId}")
-    public ResponseEntity<SlackResponseDto> readSlack(@PathVariable("slackId") UUID slackId) {
+    public ResponseEntity<SlackResponseDto> readSlack(@PathVariable("slackId") UUID slackId,
+                                                      HttpServletRequest request) {
+        String role = request.getHeader("role");
+        if (!role.equals("MASTER")) {
+            throw new ForbiddenException("마스터 권한의 유저만 접근 가능합니다.");
+        }
         SlackResponseDto responseDto = slackApplicationService.readSlack(slackId);
         return ResponseEntity.ok(responseDto);
     }
 
     @GetMapping("/search")
-    public ResponseEntity<PageResponseDto<SlackResponseDto>> searchSlack(@ModelAttribute SlackSearchDto requestDto) {
+    public ResponseEntity<PageResponseDto<SlackResponseDto>> searchSlack(@ModelAttribute SlackSearchDto requestDto,
+                                                                         HttpServletRequest request) {
+        String role = request.getHeader("role");
+        if (!role.equals("MASTER")) {
+            throw new ForbiddenException("마스터 권한의 유저만 접근 가능합니다.");
+        }
         PageResponseDto<SlackResponseDto> responseDtoPage = slackApplicationService.searchSlack(requestDto);
         return ResponseEntity.ok(responseDtoPage);
     }
 
     @PatchMapping("/{slackId}")
-    public ResponseEntity<String> deleteSlack(@PathVariable("slackId") UUID slackId) {
+    public ResponseEntity<String> deleteSlack(@PathVariable("slackId") UUID slackId,
+                                              HttpServletRequest request) {
+        String role = request.getHeader("role");
+        if (!role.equals("MASTER")) {
+            throw new ForbiddenException("마스터 권한의 유저만 접근 가능합니다.");
+        }
         slackApplicationService.deleteSlack(slackId);
         return ResponseEntity.ok("삭제 성공 : " + slackId);
     }
